% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Bacher_param_preset}
\alias{Bacher_param_preset}
\title{SPARSim simulation parameter preset from Bacher dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Bacher_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Bacher (\emph{Bacher et al., 2017; GSE85917}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Bacher count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 17128 genes
  \item 366 cells (across 4 different experimental conditions)
  \item average read/UMI count per cell of 2504380
  \item sparsity ~39\%
}
Information about the original Bacher dataset:
\itemize{
  \item Number of genes: 17128
  \item Number of cells: 366
  \item Number of different experimental conditions: 4 (Bacher_C1, Bacher_C2, Bacher_C3, Bacher_C4)
  \item Sparsity: ~39.08\%
  \item Average read/UMI count per cell: 2504380
  \item Species: Human
  \item Cell types: Embryonic stem cells
  \item Protocol and platform: Fluidigm C1, SMARTer
  \item UMIs: No
}
}
\details{
Additional details about the Bacher dataset:

scRNA-seq data were produced from H9 and H1 hESCs. 91 hESCs were captured using a Fluidigm C1 system and processed using the SMARTer protocol. The identical single-cell indexed and fragmented cDNA were pooled at 24 cells per lane or at 96 cells per lane, resulting in approximately 4 million and 1 million mapped reads per cell in the two groups, respectively. The same process was then repeated for H1 hESCs. A total of 366 cells were sequenced, resulting in:
\itemize{
  \item Bacher_C1: 91 H9 hESCs (at 24 single cell cDNA libraries per lane)
  \item Bacher_C2: 92 H1 hESCs (at 24 single cell cDNA libraries per lane)
  \item Bacher_C3: 91 H9 hESCs (at 96 single cell cDNA libraries per lane)
  \item Bacher_C4: 92 H1 hESCs (at 96 single cell cDNA libraries per lane)
}
}
\examples{
# load Bacher parameter preset
data(Bacher_param_preset)

# simulate a count table using the just loaded parameter preset
Bacher_like_dataset <- SPARSim_simulation(Bacher_param_preset)

# simulate a count table resembling only the first two experimental conditions
# i.e. Bacher_C1 and Bacher_C2
Bacher_like_dataset_2 <- SPARSim_simulation(Bacher_param_preset[c("Bacher_C1", "Bacher_C2")])

}
\references{
Bacher, R., et al. SCnorm: robust normalization of single-cell RNA-seq data. Nat Methods, 2017, 14.6: 584.
}
\keyword{datasets}
