% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Brain_10X_param_preset}
\alias{Brain_10X_param_preset}
\title{SPARSim simulation parameter preset from 10X Brain dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Brain_10X_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from 10X Brain (\emph{https://support.10xgenomics.com/single-cell-gene-expression/datasets/}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Brain_10X count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 21080 genes
  \item 11843 cells
  \item average read/UMI count per cell of 7601
  \item sparsity ~87\%
}
Information about the original 10X Brain dataset:
\itemize{
  \item Number of genes: 21080
  \item Number of cells: 11843
  \item Number of different experimental conditions: -
  \item Sparsity: ~87.37\%
  \item Average read/UMI count per cell: 7601
  \item Species: Mouse
  \item Cell types: Brain cells
  \item Protocol and platform: 10X Genomics
  \item UMIs: Yes
}
}
\details{
Additional details about the 10X Brain dataset:

scRNA-seq data were produced from brain of an E18 mouse.
Cells were processed using the Chromium Single-Cell 3′ v3 platform.
A total of 11843 cells were sequenced.
}
\examples{
# load Brain_10X parameter preset
data(Brain_10X_param_preset)

# simulate a count table using the just loaded parameter preset
Brain_10X_like_dataset <- SPARSim_simulation(Brain_10X_param_preset)

}
\references{
https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0/neuron_10k_v3
}
\keyword{datasets}
