% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Camp_param_preset}
\alias{Camp_param_preset}
\title{SPARSim simulation parameter preset from Camp dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Camp_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Camp (\emph{Camp et al., 2015; GSE75140}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Camp count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 39913 genes
  \item 434 cells (across 6 different experimental conditions)
  \item average read/UMI count per cell of 981747
  \item sparsity ~84\%
}
Information about the original Camp dataset:
\itemize{
  \item Number of genes: 39913
  \item Number of cells: 434
  \item Number of different experimental conditions: 6 (Camp_C1, Camp_C2, Camp_C3, Camp_C4, Camp_C5, Camp_C6)
  \item Sparsity: ~84.00\%
  \item Average read/UMI count per cell: 981747
  \item Species: Human
  \item Cell types: Whole brain organoids, Microdissected Cerebral Organoids
  \item Protocol and platform: Fluidigm C1, SMARTer
  \item UMIs: No
}
}
\details{
Additional details about the Camp dataset:

scRNA-seq data were produced from human fetal neocortex and human cerebral organoids.
A total of 734 cells were sequenced. All samples were processed on the microfluidic Fluidigm C1 and contain 92 external RNA spike-ins.
Samples were sequenced using the SMARTer protocol.
Cerebral organoid data were generated from dissociated whole organoids derived from induced pluripotent stem cell line 409B2 (iPSC 409B2) at 33 days (40 cells), 35 days (68 cells), 37 days (71 cells), 41 days (74 cells), and 65 days (80 cells) after the start of embryoid body culture.
Cerebral organoid data were also generated from microdissected cortical-like regions from H9 embryonic stem cell derived organoids at 53 days (region 1, 48 cells; region 2, 48 cells) or from iPSC 409B2 organoids at 58 days (region 3, 43 cells; region 4, 36 cells).
Samples from Whole brain organoids at day 41 were removed due to an error during Splat parameter estimation, resulting in 434 cells:
\itemize{
  \item Camp_C1: 96 cells from Microdissected Cerebral Organoids at day 53
  \item Camp_C2: 79 cells from Microdissected Cerebral Organoids at day 58
  \item Camp_C3: 40 cells from Whole brain organoids at day 33
  \item Camp_C4: 68 cells from Whole brain organoids at day 35
  \item Camp_C5: 71 cells from Whole brain organoids at day 37
  \item Camp_C6: 80 cells from Whole brain organoids at day 65
}
}
\examples{
# load Camp parameter preset
data(Camp_param_preset)

# simulate a count table using the just loaded parameter preset
Camp_like_dataset <- SPARSim_simulation(Camp_param_preset)

# simulate a count table resembling only the Whole brain organoids
# i.e. Camp_C3, Camp_C4, Camp_C5 and Camp_C6
Camp_like_dataset_2 <- SPARSim_simulation(Camp_param_preset[c("Camp_C3","Camp_C4",
                                                             "Camp_C5","Camp_C6")])

}
\references{
Camp, J.G., et al. Human cerebral organoids recapitulate gene expression programs of fetal neocortex development. Proc Natl Acad Sci U S A. 2015;112:15672–7.
}
\keyword{datasets}
