% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Chu_param_preset}
\alias{Chu_param_preset}
\title{SPARSim simulation parameter preset from Chu dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Chu_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Chu (\emph{Chu et al., 2016; GSE75748)}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Chu count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 17782 genes
  \item 758 cells (across 6 different experimental conditions)
  \item average read/UMI count per cell of 1335606
  \item sparsity ~51\%
}
Information about the original Chu dataset:
\itemize{
  \item Number of genes: 17782
  \item Number of cells: 758
  \item Number of different experimental conditions: 6 (Chu_C1, Chu_C2, Chu_C3, Chu_C4, Chu_C5, Chu_C6)
  \item Sparsity: ~51.30\%
  \item Average read/UMI count per cell: 1335606
  \item Species: Human
  \item Cell types: iPSC to Endoderm
  \item Protocol and platform: Fluidigm C1, SMARTer
  \item UMIs: No
}
}
\details{
Additional details about the Chu dataset:

scRNA-seq data were produced from a 6 step time series experiment of Human pluripotent stem cells (hPSCs) differentiation into definitive endoderm.
The six time points correspond to 0 h, 12 h, 24 h, 36 h, 72 h, and 96 h along the differentiation protocol from embryonic stem cell to definitive endoderm.
Single-cell capture and library preparations were performed using the Fluidigm C1 and a SMARTer protocol.
A total of 758 cells were sequenced, resulting in:
\itemize{
  \item Chu_C1: 92 cells at 0 h
  \item Chu_C2: 102 cells at 12 h
  \item Chu_C3: 66 cells at 24 h
  \item Chu_C4: 172 cells at 36 h
  \item Chu_C5: 138 cells at 72 h
  \item Chu_C6: 188 cells at 96 h
}
}
\examples{
# load Chu parameter preset
data(Chu_param_preset)

# simulate a count table using the just loaded parameter preset
Chu_like_dataset <- SPARSim_simulation(Chu_param_preset)

# simulate a count table resembling only cells at time points 0h and 96h, i.e. Chu_C1 and Chu_C6
Chu_like_dataset_2 <- SPARSim_simulation(Chu_param_preset[c("Chu_C1", "Chu_C6")])

}
\references{
Chu, L. F., et al. Single-cell RNA-seq reveals novel regulators of human embryonic stem cell differentiation to definitive endoderm. Genome Biol, 2016, 17.1: 173.
}
\keyword{datasets}
