% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Engel_param_preset}
\alias{Engel_param_preset}
\title{SPARSim simulation parameter preset from Engel dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Engel_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Engel (\emph{Engel et al., 2016; GSE74596}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Engel count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 25865 genes
  \item 203 cells (across 4 different experimental conditions)
  \item average read/UMI count per cell of 4215547
  \item sparsity ~69\%
}
Information about the original Engel dataset:
\itemize{
  \item Number of genes: 25865
  \item Number of cells: 203
  \item Number of different experimental conditions: 4 (Engel_C1, Engel_C2, Engel_C3, Engel_C4)
  \item Sparsity: ~68.59\%
  \item Average read/UMI count per cell: 4215547
  \item Species: Mouse
  \item Cell types: Natural killer T cells
  \item Protocol and platform: Flow cytometry, Modified Smart-seq2
  \item UMIs: No
}
}
\details{
Additional details about the Engel dataset:

scRNA-seq data were produced from mouse Natural killer T cells.
NKT cells subtypes (NKT0, NKT1, NKT17 and NKT2) were isolated from thymuses and directly sorted by flow cytometry into lysis buffer (96 well plate single cell sort).
A modified Smart-seq2 protocol was used for sequencing.
A total of 203 cells were sequenced, resulting in
\itemize{
  \item Engel_C1: 45 NTK0 cells
  \item Engel_C2: 46 NTK1 cells
  \item Engel_C3: 44 NTK17 cells
  \item Engel_C4: 68 NTK2 cells
}
}
\examples{
# load Engel parameter preset
data(Engel_param_preset)

# simulate a count table using the just loaded parameter preset
Engel_like_dataset <- SPARSim_simulation(Engel_param_preset)

# simulate a count table resembling only NTK1 and NTK2 cells, i.e. Engel_C2 and Engel_C4
Engel_like_dataset_2 <- SPARSim_simulation(Engel_param_preset[c("Engel_C2", "Engel_C4")])

}
\references{
Engel, I., et al., Innate-like functions of natural killer T cell subsets result from highly divergent gene programs. Nat Immunol. 2016;17:728–39.
}
\keyword{datasets}
