% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Example_count_matrix}
\alias{Example_count_matrix}
\title{Example count matrix}
\format{
A dataframe having genes on rows and cells on columns.
}
\usage{
data(Example_count_matrix)
}
\description{
Example count matrix having 5000 genes and 150 cells (across 3 different experimental conditions).
First 50 cells correspond to condition A, second 50 cells correspond to condition B and last 50 cells correspond to conditions C.
}
\examples{
# Load example count matrix
data(Example_count_matrix)

# Normalize the count matrix
Example_count_matrix_norm <- scran_normalization(Example_count_matrix)

# Get column index for each experimental condition
cond_A_column_index <- c(1:50) # Condition A column indices: from column 1 to column 50
cond_B_column_index <- c(51:100) # Condition B column indices: from column 51 to column 100
cond_C_column_index <- c(101:150) # Condition C column indices: from column 101 to column 150

# Create conditions parameter
Example_count_matrix_conditions <- list(cond_A = cond_A_column_index,
                                        cond_B = cond_B_column_index,
                                        cond_C = cond_C_column_index)

# Create SPARSim simulation parameter through the estimation from the example count matrix
SPARSim_example_sim_param <- SPARSim_estimate_parameter_from_data(raw_data = Example_count_matrix,
                                                       norm_data = Example_count_matrix_norm,
                                                       conditions = Example_count_matrix_conditions)

# Run SPARSim simulation using the just created simulation parameter
example_sim_result <- SPARSim_simulation(dataset_parameter = SPARSim_example_sim_param)

}
\keyword{datasets}
