% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_utilities.R
\name{get_rng_state}
\alias{get_rng_state}
\title{Function to retrieve the current state of the random number generation}
\usage{
get_rng_state()
}
\value{
The \code{.Random.seed} object if it exists, NULL otherwise.
}
\description{
This function return the \code{.Random.seed} object in the global environment.
}
