% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Horning_param_preset}
\alias{Horning_param_preset}
\title{SPARSim simulation parameter preset from Horning dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Horning_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Horning (\emph{Horning et al., 2018; GSE99795)}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Horning count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 37347 genes
  \item 144 cells (across 3 different experimental conditions)
  \item average read/UMI count per cell of 2572504
  \item sparsity ~39\%
}
Information about the original Horning dataset:
\itemize{
  \item Number of genes: 37347
  \item Number of cells: 144
  \item Number of different experimental conditions: 3 (Horning_C1, Horning_C2, Horning_C3)
  \item Sparsity: ~39.25\%
  \item Average read/UMI count per cell: 2572504
  \item Species: Human
  \item Cell types: LNCap (prostate cancer)
  \item Protocol and platform: Smart-seq2
  \item UMIs: No
}
}
\details{
Additional details about the Horning dataset:

scRNA-seq data were produced from LNCaP prostate cancer cells treated and untreated with androgen after cell cycle synchronization.
The three treatment groups (Group 1: cells 0h after synchronization and androgen deprivation; Group 2: cells 12h after synchronization and androgen deprivation; Group 3: cells 12h after synchronization and androgen addition) were processed using a SMART-seq2 protocol.
A total of 144 cells were sequenced, resulting in:
\itemize{
  \item Horning_C1: 48 cells of Group 1
  \item Horning_C2: 48 cells of Group 2
  \item Horning_C3: 48 cells of Group 3
}
}
\examples{
# load Horning parameter preset
data(Horning_param_preset)

# simulate a count table using the just loaded parameter preset
Horning_like_dataset <- SPARSim_simulation(Horning_param_preset)

# simulate a count table resembling only cells at time points 12h, i.e. Horning_C2 and Horning_C3
Horning_like_dataset_2 <- SPARSim_simulation(Horning_param_preset[c("Horning_C2", "Horning_C3")])

}
\references{
Horning, A.M., et al. Single-Cell RNA-seq Reveals a Subpopulation of Prostate Cancer Cells with Enhanced Cell-Cycle-Related Transcription and Attenuated Androgen Response. Cancer Res 2018 Feb 15;78(4):853-864.
}
\keyword{datasets}
