% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_utilities.R
\name{int_var_relation}
\alias{int_var_relation}
\title{Associate gene intensity and gene variability values}
\usage{
int_var_relation(intensity_values, sim_param)
}
\arguments{
\item{intensity_values}{the (new) intensity values to which we want to associate variability values}

\item{sim_param}{simulation parameter from which learn the gene intensity vs gene variability association}
}
\value{
A vector of the same size of \code{intensity_values} containing the computed variability values
}
\description{
This function associates variability values to a set of intensity values (obtained, for example, during DE genes simulation)
\code{intensity_values}.
The association between gene intensity values and gene variability values is learned from the one
described in an existing simulation parameter \code{sim_param}, using a linear interpolation.
}
