% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Macosko_param_preset}
\alias{Macosko_param_preset}
\title{SPARSim simulation parameter preset from Macosko dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Macosko_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Macosko (\emph{Macosko et al., 2015; GSE63472-GSM1626794}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Macosko count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 20230 genes
  \item 9000 cells
  \item average read/UMI count per cell of 1129
  \item sparsity ~97\%
}
Information about the original Macosko dataset:
\itemize{
  \item Number of genes: 20230
  \item Number of cells: 9000
  \item Number of different experimental conditions: -
  \item Sparsity: ~96.54\%
  \item Average read/UMI count per cell: 1129
  \item Species: Mouse
  \item Cell types: Retinal cells
  \item Protocol and platform: Drop-Seq
  \item UMIs: Yes
}
}
\details{
Additional details about the Macosko dataset:

scRNA-seq data were produced from wild-type P14 mouse retinal cells.
Cells were processed using the Drop-Seq platform.
A total of 9000 cells were sequenced.
}
\examples{
# load Macosko parameter preset
data(Macosko_param_preset)

# simulate a count table using the just loaded parameter preset
Macosko_like_dataset <- SPARSim_simulation(Macosko_param_preset)

}
\references{
Makosco, E.Z., et al. Highly Parallel Genome-wide Expression Profiling of Individual Cells Using Nanoliter Droplets. Cell. 2015:161(5):1202-1214
}
\keyword{datasets}
