% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{PBMC_10X_param_preset}
\alias{PBMC_10X_param_preset}
\title{SPARSim simulation parameter preset from 10X PBMC dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(PBMC_10X_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from 10X PBMC (\emph{https://support.10xgenomics.com/single-cell-gene-expression/datasets/}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to PBMC_10X count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 17220 genes
  \item 5419 cells
  \item average read/UMI count per cell of 3880
  \item sparsity ~96\%
}
Information about the original 10X PBMC dataset:
\itemize{
  \item Number of genes: 17220
  \item Number of cells: 5419
  \item Number of different experimental conditions: -
  \item Sparsity: ~95.66\%
  \item Average read/UMI count per cell: 3880
  \item Species: Human
  \item Cell types: Peripheral blood mononuclear cells
  \item Protocol and platform: 10X Genomics
  \item UMIs: Yes
}
}
\details{
Additional details about the 10X PBMC dataset:

scRNA-seq data were produced from peripheral blood mononuclear cells (PBMCs) from a healthy donor.
Cells were processed using the Chromium Single-Cell 3′ v1 platform.
A total of 5419 cells were sequenced.
}
\examples{
# load PBMC_10X parameter preset
data(PBMC_10X_param_preset)

# simulate a count table using the just loaded parameter preset
PBMC_10X_like_dataset <- SPARSim_simulation(PBMC_10X_param_preset)

}
\references{
https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc6k
}
\keyword{datasets}
