% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_simulate.R
\name{runif_int_no_duplicates}
\alias{runif_int_no_duplicates}
\title{Generate integer number from an uniform distribution with no duplicates}
\usage{
runif_int_no_duplicates(num, max_val, seed)
}
\arguments{
\item{num}{sample size of the MH}

\item{max_val}{max value for random number generation in the range \code{[1, max_val]}}

\item{seed}{numeric seed}
}
\value{
An array of \code{length(avgAbund)} elements representing the count values for the current sample
}
\description{
Function to generate \code{num} integer values from an uniform distribution \code{[1, max_val]} with no duplicates
with control on random number generation through seed.
}
