% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Saunders_param_preset}
\alias{Saunders_param_preset}
\title{SPARSim simulation parameter preset from Saunders dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Saunders_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Saunders (\emph{Saunders et al., 2018; GSE116470}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Saunders count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 19940 genes
  \item 5688 cells
  \item average read/UMI count per cell of 2183
  \item sparsity ~94\%
}
Information about the original Saunders dataset:
\itemize{
  \item Number of genes: 19940
  \item Number of cells: 5688
  \item Number of different experimental conditions: -
  \item Sparsity: ~94.10\%
  \item Average read/UMI count per cell: 2183
  \item Species: Mouse
  \item Cell types: Polydendrocytes
  \item Protocol and platform: Drop-Seq
  \item UMIs: Yes
}
}
\details{
Additional details about the Saunders dataset:

scRNA-seq data were produced from P60-70 C57BL/6 male mice.
Cells were processed using the Drop-Seq platform.
A total of 5688 cells were sequenced.
}
\examples{
# load Saunders parameter preset
data(Saunders_param_preset)

# simulate a count table using the just loaded parameter preset
Saunders_like_dataset <- SPARSim_simulation(Saunders_param_preset)

}
\references{
Saunders, A., et al. Molecular Diversity and Specializations among the Cells of the Adult Mouse Brain. Cell. 2018:174(4) P1015-1030.E16

http://dropviz.org/
}
\keyword{datasets}
