% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_utilities.R
\name{scran_normalization}
\alias{scran_normalization}
\title{scran normalization}
\usage{
scran_normalization(data, sizes = seq(20, 100, 5), positive = FALSE)
}
\arguments{
\item{data}{a raw count matrix (genes on rows, samples on columns) to normalize}

\item{sizes}{parameter \code{size} of function \code{scran::computeSumFactors}}

\item{positive}{parameter \code{positive} of function \code{scran::computeSumFactors}}
}
\value{
The normalized count matrix using the scran method
}
\description{
Function to compute the count normalization proposed by Lun et. al in the scran package
}
