% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_utilities.R
\name{set_rng_state}
\alias{set_rng_state}
\title{Function to restore the a (previously saved) state of the random number generation}
\usage{
set_rng_state(rng_state)
}
\arguments{
\item{rng_state}{A (previously saved) state of the random number generation}
}
\description{
This function restore the \code{.Random.seed} object in the global environment with the one passed as input argument.
If the input argument is NULL, then no action is performed.
}
