% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-package.R
\docType{package}
\name{SPARSim}
\alias{SPARSim}
\title{SPARSim single cell: a count data simulator for scRNA-seq data}
\description{
This package implements the SPARSim simulator, a tool to simulate single cell RNA-seq count table.
For additional details about the simulation model, please refer to
\emph{Baruzzo, G., Patuzzi, I., & Di Camillo, B. (2020). SPARSim single cell: a count data simulator for scRNA-seq data. Bioinformatics, 36(5), 1468-1475.}
}
\references{
Baruzzo, G., Patuzzi, I., & Di Camillo, B. (2020). SPARSim single cell: a count data simulator for scRNA-seq data. Bioinformatics, 36(5), 1468-1475.
}
\author{
Giacomo Baruzzo \email{giacomo.baruzzo@unipd.it}

Ilaria Patuzzi \email{IPatuzzi@izsvenezie.it}

Barbara Di Camillo \email{barbara.dicamillo@unipd.it}
}
