% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_batch_effect.R
\name{SPARSim_compute_batch_effect_factors}
\alias{SPARSim_compute_batch_effect_factors}
\title{Generate batch effect multiplicative factors for each batch}
\usage{
SPARSim_compute_batch_effect_factors(
  N_batch_factors,
  batch_factors_id,
  SPARSim_batch_parameter
)
}
\arguments{
\item{N_batch_factors}{number of features that will be affected by batch effect multiplicative factors, i.e. number of features in the count matrix}

\item{batch_factors_id}{IDs of features that will be affected by batch effect multiplicative factors, i.e. names of the features in the count matrix}

\item{SPARSim_batch_parameter}{SPARSim batch parameter containing the information about the batch effect}
}
\value{
An updated version of the input parameter \code{SPARSim_batch_parameter} containing the values of the multiplicative factors for each batch
}
\description{
Function to generate the batch effect multiplicative factors for each batch, following the batch description
in \code{SPARSim_batch_parameter}.
}
