% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_batch_effect.R
\name{SPARSim_create_batch}
\alias{SPARSim_create_batch}
\title{Create a SPARSim batch object}
\usage{
SPARSim_create_batch(name, distribution = "normal", param_A = 1, param_B = 1)
}
\arguments{
\item{name}{name of the batch}

\item{distribution}{batch effect factors distribution. current available distributions:
\itemize{
  \item "normal": normal distribution (mean = param_A, sd = param_B)
  \item "gamma": gamma distribution (scale = param_A, shape = param_B)
}}

\item{param_A}{parameter to describe the distribution}

\item{param_B}{parameter to describe the distribution}
}
\value{
A SPARSim batch object
}
\description{
Function to create a SPARSim batch object.
}
\details{
A SPARSim batch object describe the characteristics of the batch effect to apply during SPARSim simulation.
Input parameter \code{name} contains a batch identifier (e.g. "Lane_1", "Lane_2", etc.).
Batch effect is simulated multiplying gene expression levels by batch specific multiplicative factors.
Batch specific multiplicative factors will be sampled from the distribution specified in the
input parameter \code{distribution}, with parameter \code{param_A} and \code{param_B}.
Currently, distribution could be one between \code{"normal"} (i.e. \code{rnorm(..., mean = param_A, sd = param_B)})
and \code{"gamma"} (i.e. \code{rgamma(..., scale = param_A, shape = param_B)})
}
\examples{
# create 2 batches, called "Lane1" and "Lane2"
batch_1 <- SPARSim_create_batch(name = "Lane1", distribution = "normal", param_A = 1, param_B = 1)
batch_2 <- SPARSim_create_batch(name = "Lane2", distribution = "normal", param_A = 3, param_B = 1)

}
