% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_batch_effect.R
\name{SPARSim_create_batch_effect_matrix}
\alias{SPARSim_create_batch_effect_matrix}
\title{Create batch effect multiplicative factors matrix}
\usage{
SPARSim_create_batch_effect_matrix(
  batch_factor_matrix,
  SPARSim_batch_parameter
)
}
\arguments{
\item{batch_factor_matrix}{matrix to fill with batch factor values (features on rows; samples on columns)}

\item{SPARSim_batch_parameter}{SPARSim batch parameter containing the information about the batch effect}
}
\value{
A numeric matrix, having the same dimensions and names of \code{batch_factor_matrix}.
The element \code{[i,j]} in the matrix contains the batch effect factors for i-th feature in j-th sample.
}
\description{
Function to create batch effects multiplicative factors matrix. This function is used internally by \code{\link{SPARSim_simulation}}.
}
\details{
This function return a matrix containing the batch effect multiplicative factors.
The matrix has the same dimensions and names of \code{batch_factor_matrix}.
The element \code{[i,j]} in the matrix contains the batch effect factor for i-th feature in j-th sample.
}
