% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_batch_effect.R
\name{SPARSim_create_batch_parameter}
\alias{SPARSim_create_batch_parameter}
\title{Create SPARSim batch effect simulation parameter}
\usage{
SPARSim_create_batch_parameter(batch_set, batch_sample)
}
\arguments{
\item{batch_set}{result of function \code{\link{SPARSim_create_batch_set}}}

\item{batch_sample}{a character array of length N, where N is the number of cells/samples to simulate.
The i-th element in the array contains the name of the batch type to simulate in the i-th sample;
if the i-th element is set to NULL, then the i-th sample is affected by no batch effect}
}
\value{
A SPARSim batch effect simulation parameter
}
\description{
Function to create the SPARSim input batch parameter.
}
\examples{
# create 2 batches, called "Lane1" and "Lane2"
batch_1 <- SPARSim_create_batch(name = "Lane1", distribution = "normal", param_A = 1, param_B = 1)
batch_2 <- SPARSim_create_batch(name = "Lane2", distribution = "normal", param_A = 3, param_B = 1)

# create batch set
batches <- list(); batches[[1]] <- batch_1; batches[[2]] <- batch_2
batch_set <- SPARSim_create_batch_set(batch_list = batches)

# create the association between batches and samples/cells
# as example, consider simulating 40 cells:
# the first 30 samples are affected by batch "Lane1",
# the second 10 samples are affected by batch "Lane2"
batch_sample <- c(rep("Lane1", 30),rep("Lane2", 10))

# having both the batch set (i.e. batch_set) and
# the association between batches and samples/cells (i.e. batch_sample)
# it is now possible to create the SPARSim batch effect simulation parameter
SPARSim_batch_parameter <- SPARSim_create_batch_parameter(batch_set = batch_set,
                                                          batch_sample = batch_sample)

}
