% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_batch_effect.R
\name{SPARSim_create_batch_set}
\alias{SPARSim_create_batch_set}
\title{Create a SPARSim batch set}
\usage{
SPARSim_create_batch_set(batch_list)
}
\arguments{
\item{batch_list}{a list of SPARSim batch objects. Each element of the list must be the result of a call to function \code{\link{SPARSim_create_batch}}}
}
\value{
A SPARSim batch set
}
\description{
Function to create a set of SPARSim batches
}
\examples{
# create 2 batches, called "Lane1" and "Lane2"
batch_1 <- SPARSim_create_batch(name = "Lane1", distribution = "normal", param_A = 1, param_B = 1)
batch_2 <- SPARSim_create_batch(name = "Lane2", distribution = "normal", param_A = 3, param_B = 1)

# create batch set
batches <- list(); batches[[1]] <- batch_1; batches[[2]] <- batch_2
batch_set <- SPARSim_create_batch_set(batch_list = batches)

}
