% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_parameter.R
\name{SPARSim_create_DE_genes_parameter}
\alias{SPARSim_create_DE_genes_parameter}
\title{Create SPARSim simulation parameter with DE genes}
\usage{
SPARSim_create_DE_genes_parameter(
  sim_param,
  fc_multiplier,
  N_cells = NULL,
  lib_size_DE = NULL,
  sample_names = NULL,
  condition_name = NULL
)
}
\arguments{
\item{sim_param}{SPARSim simulation parameter used as prototype}

\item{fc_multiplier}{Array of fold change multipliers}

\item{N_cells}{Number of cells to simulate.
If NULL (default), the number of cells will be set to the length of \code{lib_size_DE} or to the number of cells in \code{sim_param}}

\item{lib_size_DE}{Array of library size values. It must be of the same length of \code{sample_names} and equal to \code{N_cells}, if they are specified.
If NULL (default), and \code{N_cells} is NULL, the library size values from sim_param will be used.
If NULL (default), and \code{N_cells} is specified, then N_cell randomly chosen library size values from sim_param will be used.}

\item{sample_names}{Array of sample names. It must be of the same length of \code{lib_size_DE} and equal to \code{N_cells}, if they are specified.
If NULL (defatul), sample will be automatically named "<condition_name>_cell1", "<condition_name>_cell2", ...}

\item{condition_name}{Name associated to the experimental condition.
If NULL (default), the name is set to the concatenation of the name in \code{sim_param} with the suffix \code{"_DE"}}
}
\value{
SPARSim simulation parameter with DE genes respect to the input SPARSim simulation parameter \code{sim_param}
}
\description{
Function to create a SPARSim simulation parameter with DE genes respect to another SPARSim simulation parameter provided as input.
}
\details{
To simulate N feature (e.g. genes), user must specify N values of gene expression level and gene expression variability in the function input parameters \code{intensity} and \code{variability}, respectively.
To simulate M samples (i.e. cells), user must specify M values of sample library size in the function input parameter \code{library_size}.

User can optionally specify the names to assign at the single feature and sample to simulate (function input parameters \code{feature_names} and \code{sample_names}, respectively,
as well as the name of the experimental condition (function input parameter \code{condition_name}). If the user does not specify such information, the function will set some default values.

To simulate T different experimental conditions in a single count table, then T different simulation parameters must be created.
}
