% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_spikein.R
\name{SPARSim_create_spikein_mix}
\alias{SPARSim_create_spikein_mix}
\title{Create a spike-in mix}
\usage{
SPARSim_create_spikein_mix(
  mix_name,
  abundance,
  spike_in_IDS = NA,
  extra_variability = NA
)
}
\arguments{
\item{mix_name}{name of the spike-in mix}

\item{abundance}{numeric array, the i-th element contains the abundance of the i-th spike-in.}

\item{spike_in_IDS}{optional array of spike-in IDs; if NA (default) spike-ins will be named "spikein_1", "spikein_2", ...,}

\item{extra_variability}{optional array of extra variability values; if NA (default), no extra variability is added during spike-ins simulation}
}
\value{
A spike-in mix object
}
\description{
Function to create a spike-in mix
}
\details{
A spike-in mix is a set of spike-ins. Each spike-in mix is identified by a name (input parameter \code{name}).

The abundances of spike-ins in the mix are described in the parameter \code{abundance}, a numeric vector containing
the abundance of each spike-in in the spike-in mix.
The length of \code{abundance} defines the number of spike-in in the spike-in mix.

By default, spike-ins are named "spikein_1", "spikein_2", ..., . The user could specified different spike-in IDs using
the parameter \code{spike_in_IDS}, which should contain an array of spike-in IDs having the same number of element of \code{abundance}.

Spike-ins should be affected only by technical variability, since their are synthetic sequences added at kwown concentration.
However, in some studies an extra variability was detected.
Such extra variability could be optionally added using the \code{extra_variability} input parameter, having the same meaning of the \code{variability}
parameter used in function \code{\link{SPARSim_create_simulation_parameter}}.
By default, no extra variability is simulated.
}
\examples{
# create a first spike-in mix made by 100 spike-ins. Let's call it "spikein_M1"
# and set spike-in IDs to "spikein_M1_1", "spikein_M1_2", ..., "spikein_M1_100"
spikein_mix1_abundance <- runif(n = 100, 0.01, 1000)
spikein_mix1_IDS <- paste0("spikein_M1_", c(1:100))
spikein_mix1 <- SPARSim_create_spikein_mix(mix_name= "spikein_M1",
                                           abundance = spikein_mix1_abundance,
                                           spike_in_IDS = spikein_mix1_IDS)

# create a second spike-in mix made by 90 spike-ins
# let's call it "spikein_M2" and add some extra variability to it
spikein_mix2_abundance <- runif(n = 90, 0.001, 10000)
spikein_mix2_extra_var <- runif(n = 90, 0.01, 0.02)
spikein_mix2 <- SPARSim_create_spikein_mix(mix_name= "spikein_M2",
                                           abundance = spikein_mix2_abundance,
                                           extra_variability = spikein_mix2_extra_var)

}
