% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_spikein.R
\name{SPARSim_create_spikein_parameter}
\alias{SPARSim_create_spikein_parameter}
\title{Create SPARSim spike-in simulation parameter}
\usage{
SPARSim_create_spikein_parameter(
  spikein_set,
  spikein_sample,
  spikein_proportion
)
}
\arguments{
\item{spikein_set}{a set of spike-in mixes, as result of a call to function \code{\link{SPARSim_create_spikein_set}}}

\item{spikein_sample}{a character array of length equal to the number of samples.
The i-th element in the array contains the name of the spike-in mix to simulate in the i-th sample;
if the i-th element is set to NULL, then the i-th sample will contain no spike-in mix.}

\item{spikein_proportion}{a numeric array of length \code{length(spikein_set)}, the i-th element in the array is associated to the i-th spike-in mix in \code{spikein_set}.
Given a reference amount of endogenous material X (here we'll set X to the median of the amount of endogenous material among all the input samples in all the experimental conditions),
the i-th element in the array represents the proportion (i.e. a value in [0,1]) of the i-th spike-in mix material respect to the reference amount X.
As example, if \code{spikein_set} contains 2 spike-in mixes, e.g. "spikein_M1" and "spikein_M2", having spikein_proportion set to \code{c(0.03,0.05)}
and having the reference amount of endogenous material X = 100ng, then:
\itemize{
  \item calling Y1 the simulated amount of material from spike-in mix "spikein_M1", Y1 = 0.03*100ng = 3ng of spike-in mix "spikein_M1" will be add to samples which require spike-in mix "spikein_M1"
  \item calling Y2 the simulated amount of material from spike-in mix "spikein_M2", Y2 = 0.05*100ng = 5ng of spike-in mix "spikein_M2" will be add to samples which require spike-in mix "spikein_M2"
}}
}
\value{
A SPARSim spike-in simulation parameter
}
\description{
Function to create a SPARSim spike-in simulation parameter
}
\examples{
# create a first spike-in mix made by 100 spike-ins. Let's call it "spikein_M1"
# and set spike-in IDs to "spikein_M1_1", "spikein_M1_2", ..., "spikein_M1_100"
spikein_mix1_abundance <- runif(n = 100, 0.01, 1000)
spikein_mix1_IDS <- paste0("spikein_M1_", c(1:100))
spikein_mix1 <- SPARSim_create_spikein_mix(mix_name= "spikein_M1",
                                           abundance = spikein_mix1_abundance,
                                           spike_in_IDS = spikein_mix1_IDS)

# create a second spike-in mix made by 90 spike-ins
# let's call it "spikein_M2" and add some extra variability to it
spikein_mix2_abundance <- runif(n = 90, 0.001, 10000)
spikein_mix2_extra_var <- runif(n = 90, 0.01, 0.02)
spikein_mix2 <- SPARSim_create_spikein_mix(mix_name= "spikein_M2",
                                           abundance = spikein_mix2_abundance,
                                           extra_variability = spikein_mix2_extra_var)

# create set of spike-in mixes
spikein_mixes_list <- list(mix1 = spikein_mix1, mix2 = spikein_mix2)
spikein_set <- SPARSim_create_spikein_set(spikein_mixes = spikein_mixes_list)

# create the association between samples/cells and spike-in mixes
# as an example, consider having 40 samples
# the first 10 samples will contain spike-in mix "spikein_M1",
# the following 10 samples will contain no spike-in mix
# and the last 20 samples will contain spike-in mix "spikein_M2"
spikein_sample <- c(rep("spikein_M1", 10), rep(NA, 10), rep("spikein_M2", 20))

# define the proportion of the spike-in mixes respect to the endogenous material
spikein_proportion <- c(0.03,0.05)

# having the set of spike-in mixes, the association between samples
# and spike-in mixes and the proportion of spike-ins abundances respect to the endogenous material,
# it is now possible to create the SPARSim spike-in simulation parameter
SPARSim_spikein_param <- SPARSim_create_spikein_parameter(spikein_set = spikein_set,
                                                          spikein_sample = spikein_sample,
                                                          spikein_proportion = spikein_proportion)

}
