% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_spikein.R
\name{SPARSim_create_spikein_set}
\alias{SPARSim_create_spikein_set}
\title{Create a set of spike-in mixes}
\usage{
SPARSim_create_spikein_set(spikein_mixes)
}
\arguments{
\item{spikein_mixes}{a list of existing spike-in mixes. Each element of the list must be the result of a call to function \code{\link{SPARSim_create_spikein_mix}}}
}
\value{
A set of spike-in mixes
}
\description{
Function to create a pool of spike-in mixes
}
\details{
If different spike-in mixes have spike-ins with the same names, such spike-ins are considered to be the same spike-ins, but maintaining their original values.

As a toy example, consider a spike-in mix A made by 5 spike-ins "spike_1", "spike_2", "spike_3", "spike_4", and "spike_5" with abundances 10, 20, 30, 40 and 50, respectively.
Consider a spike-in mix B made by 4 spike-ins "spike_4", "spike_5", "spike_6" and "spike_7", with abundances 400, 500, 600 and 700 respectively.

The set of spike-in mixes is so made by 7 spike-ins: "spike_1", "spike_2", "spike_3", "spike_4", "spike_5", "spike_6" and "spike_7".
In this scenario, "spike_4" and "spike_5" are common to spike-in mix A and mix B.

However, they will maintain their original values: "spike_4" and "spike_5" will have abundances of 40 and 50 when associated to spike-in mix A,
while they will have abundances of 400 and 500 when associated to spike-in mix B.
}
\examples{
# create a first spike-in mix made by 100 spike-ins. Let's call it "spikein_M1"
# and set spike-in IDs to "spikein_M1_1", "spikein_M1_2", ..., "spikein_M1_100"
spikein_mix1_abundance <- runif(n = 100, 0.01, 1000)
spikein_mix1_IDS <- paste0("spikein_M1_", c(1:100))
spikein_mix1 <- SPARSim_create_spikein_mix(mix_name= "spikein_M1",
                                           abundance = spikein_mix1_abundance,
                                           spike_in_IDS = spikein_mix1_IDS)

# create a second spike-in mix made by 90 spike-ins
# let's call it "spikein_M2" and add some extra variability to it
spikein_mix2_abundance <- runif(n = 90, 0.001, 10000)
spikein_mix2_extra_var <- runif(n = 90, 0.01, 0.02)
spikein_mix2 <- SPARSim_create_spikein_mix(mix_name= "spikein_M2",
                                           abundance = spikein_mix2_abundance,
                                           extra_variability = spikein_mix2_extra_var)

# create set of spike-in mixes
spikein_mixes_list <- list(mix1 = spikein_mix1, mix2 = spikein_mix2)
spikein_set <- SPARSim_create_spikein_set(spikein_mixes = spikein_mixes_list)

}
