% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_estimate.R
\name{SPARSim_estimate_parameter_from_data}
\alias{SPARSim_estimate_parameter_from_data}
\title{Estimate SPARSim simulation parameter from a given count table}
\usage{
SPARSim_estimate_parameter_from_data(raw_data, norm_data, conditions)
}
\arguments{
\item{raw_data}{count matrix (gene on rows, samples on columns) containing raw count data}

\item{norm_data}{count matrix (gene on rows, samples on columns) containing normalized count data}

\item{conditions}{list where the i-th element of the list contains the column indices for i-th experimental conditions. List must be a named list.}
}
\value{
A SPARSim simulation parameters
}
\description{
Function to estimate SPARSim simulation parameters (intensity, variability and library sizes) from a real count table.
If the real count table contains more than one experimental condition, it is possible to estimate the parameters for each experimental condition.
}
