% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim_simulate.R
\name{SPARSim_simulation}
\alias{SPARSim_simulation}
\title{Function to simulate a raw count table}
\usage{
SPARSim_simulation(
  dataset_parameter,
  batch_parameter = NULL,
  spikein_parameter = NULL,
  output_sim_param_matrices = FALSE,
  output_batch_matrix = FALSE,
  gene_expr_simulation_seed = NULL,
  count_data_simulation_seed = NULL,
  bimodal_expr_simulation_seed = NULL,
  batch_effect_simulation_seed = NULL,
  preserve_global_rng = FALSE
)
}
\arguments{
\item{dataset_parameter}{list containing, the intensity, variability and lib sizes of each experimental condition. It is the return value of "estimate_parameter_from_data" or could be created by the users}

\item{batch_parameter}{batch effect simulation parameter}

\item{spikein_parameter}{spike-in simulation parameter}

\item{output_sim_param_matrices}{boolean flag. If TRUE, the function will output two additional matrices, called abundance_matrix and variability_matrix, containing the gene intensities and gene variabilities used as simulation input. (Default: FALSE)}

\item{output_batch_matrix}{boolean flag. If TRUE, the function will output an additional matrix, called batch_factors_matrix, containing the multiplicative factors used in batch effect simulation. (Default: FALSE)}

\item{gene_expr_simulation_seed}{integer value used as seed for random number generation during gene expression level simulation. If NULL (default) the current R seed is used.}

\item{count_data_simulation_seed}{integer value used as seed for random number generation during count data simulation. If NULL (default) the current R and C++ seed is used.}

\item{bimodal_expr_simulation_seed}{integer value used as seed for random number generation during bimodal gene expression level simulation. If NULL (default) the current R seed is used.}

\item{batch_effect_simulation_seed}{integer value used as seed for random number generation during batch effect simulation. If NULL (default) the current R seed is used.}

\item{preserve_global_rng}{boolean flag. If TRUE, the function will not alter the state of the random number generator in the global R environment. If FALSE (default), the state of the random number generator may be altered during the execution of the function.}
}
\value{
A list of 5 elements:

  - \code{count_matrix}: the simulated count matrix (genes on rows, samples on columns)

  - \code{gene_matrix}: the simulated gene expression levels (genes on rows, samples on columns)

  - \code{abundance_matrix}: the input gene intensity values provided as input (genes on rows, samples on columns), if \code{output_sim_param_matrices} = TRUE. NULL otherwise.

  - \code{variability_matrix}: the input gene variability values provided as input (genes on rows, samples on columns), if \code{output_sim_param_matrices} = TRUE. NULL otherwise.

  - \code{batch_factors_matrix}: the multiplicative factor used in batch generation (genes on rows, samples on columns), if \code{output_batch_matrix} = TRUE. NULL otherwise.
}
\description{
Function to simulate a raw count table
}
