% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{T_10X_param_preset}
\alias{T_10X_param_preset}
\title{SPARSim simulation parameter preset from 10X T dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(T_10X_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from 10X T (\emph{https://support.10xgenomics.com/single-cell-gene-expression/datasets/}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to T_10X count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 20267 genes
  \item 8093 cells
  \item average read/UMI count per cell of 3880
  \item sparsity ~95\%
}
Information about the original 10X T dataset:
\itemize{
  \item Number of genes: 20267
  \item Number of cells: 8093
  \item Number of different experimental conditions: -
  \item Sparsity: ~94.59\%
  \item Average read/UMI count per cell: 3880
  \item Species: Human
  \item Cell types: Pan T cells
  \item Protocol and platform: 10X Genomics
  \item UMIs: Yes
}
}
\details{
Additional details about the 10X T dataset:

scRNA-seq data were produced from pan T cells isolated from mononuclear cells of a healthy donor.
Cells were processed using the Chromium Single-Cell 3′ v2 platform.
A total of 8093 cells were sequenced.
}
\examples{
# load T_10X parameter preset
data(T_10X_param_preset)

# simulate a count table using the just loaded parameter preset
T_10X_like_dataset <- SPARSim_simulation(T_10X_param_preset)

}
\references{
https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/t_3k_4k_aggregate
}
\keyword{datasets}
