% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Tung_param_preset}
\alias{Tung_param_preset}
\title{SPARSim simulation parameter preset from Tung dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Tung_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Tung (\emph{Tung et al., 2017; GSE77288}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Tung count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 18464 genes
  \item 564 cells (across 8 different experimental conditions)
  \item average read/UMI count per cell of 70473
  \item sparsity ~53\%
}
Information about the original Tung dataset:
\itemize{
  \item Number of genes: 18464
  \item Number of cells: 564
  \item Number of different experimental conditions: 8 (Tung_C1, Tung_C2, Tung_C3, Tung_C4, Tung_C5, Tung_C6, Tung_C7, Tung_C8)
  \item Sparsity: ~53.16\%
  \item Average read/UMI count per cell: 70473
  \item Species: Human
  \item Cell types: Induced pluripotent stem cells
  \item Protocol and platform: Fluidigm C1, modified SMARTer
  \item UMIs: Yes
}
}
\details{
Additional details about the Tung dataset:

scRNA-seq data were produced from induced pluripotent stem cell (iPSC) lines of three Yoruba individuals (NA19098, NA19101 and NA19239).
Three C1 96 well-integrated fluidic circuit replicates were collected from each individual and ERCC spike-in control were added.
Sample were sequenced using the SMARTer protocol, using 5-bp UMIs. A total of 864 cells were sequenced.
Cells from second replicate of individual NA19098 were removed due to a pipetting error during in ERCC addition.
Additional cells were removed using an ad-hoc quality control procedure (see (\emph{Tung et al., 2017})).
Finally, only 564 cells were marked as high quality samples, resulting in:
\itemize{
  \item Tung_C1: 85 cells from first replicate of NA19098
  \item Tung_C2: 57 cells from third replicate of NA19098
  \item Tung_C3: 80 cells from first replicate of NA19101
  \item Tung_C4: 70 cells from second replicate of NA19101
  \item Tung_C5: 51 cells from third replicate of NA19101
  \item Tung_C6: 74 cells from first replicate of NA19239
  \item Tung_C7: 68 cells from second replicate of NA19239
  \item Tung_C8: 79 cells from third replicate of NA19239
}
}
\examples{
# load Tung parameter preset
data(Tung_param_preset)

# simulate a count table using the just loaded parameter preset
Tung_like_dataset <- SPARSim_simulation(Tung_param_preset)

# simulate a count table resembling only the last three experimental conditions
# i.e. Tung_C6, Tung_C7 and Tung_C8
Tung_like_dataset_2 <- SPARSim_simulation(Tung_param_preset[c("Tung_C6", "Tung_C7", "Tung_C8")])

}
\references{
Tung, P.Y. et al. Batch effects and the effective design of single-cell gene expression studies. Sci Rep. 2017;7:39921.
}
\keyword{datasets}
