% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARSim-data.R
\docType{data}
\name{Zheng_param_preset}
\alias{Zheng_param_preset}
\title{SPARSim simulation parameter preset from Zheng dataset}
\format{
A list of one or more SPARSim simulation parameters (one of each experimental condition) as output of function \code{\link{SPARSim_estimate_parameter_from_data}}.
The single simulation parameter has the format described in \code{\link{SPARSim_create_simulation_parameter}}.
}
\usage{
data(Zheng_param_preset)
}
\description{
SPARSim simulation parameter preset estimated from Zheng (\emph{Zheng et al., 2017}) count table.
It allows simulating a count table having characteristics (i.e. number of genes, number of cells, gene intensity, gene variability, sparsity and library sizes) similar to Zheng count matrix.
The simulation parameter will generate a count matrix having:
\itemize{
  \item 19536 genes
  \item 3388 cells (across 4 different experimental conditions)
  \item average read/UMI count per cell of 15041
  \item sparsity ~83\%
}
Information about the original Zheng dataset:
\itemize{
  \item Number of genes: 19536
  \item Number of cells: 3388
  \item Number of different experimental conditions: 4 (Zheng_C1, Zheng_C2, Zheng_C3, Zheng_C4)
  \item Sparsity: ~82.56\%
  \item Average read/UMI count per cell: 15041
  \item Species: Human
  \item Cell types: Jurkat and 293T cells
  \item Protocol and platform: 10X Genomics
  \item UMIs: Yes
}
}
\details{
Additional details about the Zheng dataset:

scRNA-seq data were produced from a sample where an equal number of 293T and Jurkat cells was mixed.
Cells were processed using the GemCode Single-Cell 3′ platform.
A total of 3388 cells were sequenced, resulting in:
\itemize{
  \item Zheng_C1: 1440 Jurkat cells
  \item Zheng_C2: 1718 293T cells
  \item Zheng_C3: 184 mixed cells, i.e. samples containing both Jurkat and 293T cells
  \item Zheng_C4: 46 unclassified cells
}
}
\examples{
# load Zheng parameter preset
data(Zheng_param_preset)

# simulate a count table using the just loaded parameter preset
Zheng_like_dataset <- SPARSim_simulation(Zheng_param_preset)

# simulate a count table resembling only Jurkat and 293T cells, i.e. Zheng_C1 and Zheng_C2
Zheng_like_dataset_2 <- SPARSim_simulation(Zheng_param_preset[c("Zheng_C1", "Zheng_C2")])

}
\references{
Zheng, G.H.Y., et al. Massively parallel digital transcriptional profiling of single cells. Nat Commun. 2017; 8:14049

https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/jurkat:293t_50:50
}
\keyword{datasets}
